requirements.txt
streamlit
gdown
pdfminer.six
python-docx
langchain
openai
faiss-cpu
sentence-transformers
huggingface-hub
spacy
langchain-community
PyPDF2